requirements_dev.txt
homeassistant
pyserial
